move dscp api file
