+++ app/server.js
@@ -14,6 +14,7 @@ const v1ApiDoc = require('./api-v1/api-doc')
 const v1RecipeService = require('./api-v1/services/recipeService')
 const v1AttachmentService = require('./api-v1/services/attachmentService')
 const v1BuildService = require('./api-v1/services/buildService')
+const v1DscpApiService = require('./api-v1/services/dscpApiService')
 const v1OrderService = require('./api-v1/services/orderService')
 const v1PartService = require('./api-v1/services/partService')
 const v1IdentityService = require('./api-v1/services/identityService')
@@ -61,6 +62,7 @@ async function createHttpServer() {
       recipeService: v1RecipeService,
       attachmentService: v1AttachmentService,
       buildService: v1BuildService,
+      dscpApiService: v1DscpApiService,
       partService: v1PartService,
       orderService: v1OrderService,
       identityService: v1IdentityService,
